first build works

+++ SlideCrafting.Utils/OS.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Text;
+using System.Threading;
 using System.Threading.Tasks;
 using YamlDotNet.RepresentationModel;
 
@@ -33,49 +34,79 @@ namespace SlideCrafting.Utils
             }
         }
 
-        public static void CreateFolder(string distFolder)
+        public static void CreateFolder(string distFolder, CancellationToken token)
         {
+            if (token.IsCancellationRequested)
+            {
+                throw new TaskCanceledException("Create folder request on already canceled task");
+            }
+
             Directory.CreateDirectory(distFolder);
         }
 
-        public static void MoveAllFilesFromFolderToFolder(string fromFolder, string toFolder, bool shouldRethrow = false) {
+        public static void MoveAllFilesFromFolderToFolder(string fromFolder, string toFolder, CancellationToken token, bool shouldRethrow = false)
+        {
             foreach (var file in Directory.EnumerateFiles(fromFolder))
             {
+                if (token.IsCancellationRequested)
+                {
+                    throw new TaskCanceledException("move all files request on already canceled task");
+                }
+
                 try
                 {
                     File.Move(file, Path.Combine(toFolder, Path.GetFileName(file)), overwrite: true);
                 }
-                catch
+                catch (Exception exc)
                 {
-                    try
+                    if (shouldRethrow)
                     {
-                        File.Copy(file, Path.Combine(toFolder, Path.GetFileName(file)), overwrite: true);
-                    }
-                    catch (Exception exc)
-                    {
-                        if (shouldRethrow)
-                        {
-                            throw new FileLoadException("can not load file to move or copy (" + file + ")", file, exc);
-                        }
+                        throw new FileLoadException("can not load file to move (" + file + ")", file, exc);
                     }
                 }
             }
 
-            foreach (var sourceFolder in Directory.EnumerateDirectories((fromFolder)))
+            foreach (var sourceFolder in Directory.EnumerateDirectories((fromFolder)).ToList())
             {
-                if (!sourceFolder.StartsWith(toFolder))
+                if (sourceFolder.TrimEnd(Path.DirectorySeparatorChar).EndsWith(".git")) 
                 {
-                    var dirInfo = new DirectoryInfo(sourceFolder);
-                    var targetFolder = Path.Combine(toFolder, dirInfo.Name);
-                    CreateFolder(targetFolder);
-                    MoveAllFilesFromFolderToFolder(sourceFolder, targetFolder);
-                    RemoveFolder(sourceFolder);
+                    continue;
+                }
+
+                if (token.IsCancellationRequested)
+                {
+                    throw new TaskCanceledException("move all files request on already canceled task");
+                }
+
+                try
+                {
+                    if(string.Compare(
+                        Path.GetFullPath(sourceFolder).TrimEnd(Path.DirectorySeparatorChar),
+                        Path.GetFullPath(toFolder).TrimEnd(Path.DirectorySeparatorChar),
+                        StringComparison.InvariantCultureIgnoreCase) != 0)
+                    {
+                        var dirInfo = new DirectoryInfo(sourceFolder);
+                        CopyFolder(sourceFolder, toFolder, token, true);
+                        RemoveFolder(sourceFolder, token);
+                    }
+                }
+                catch (Exception exc)
+                {
+                    if (shouldRethrow)
+                    {
+                        throw new FileLoadException("can not load dir to move (" + sourceFolder + ")", sourceFolder, exc);
+                    }
                 }
             }
         }
 
-        public static void RemoveFolder(string folder)
+        public static void RemoveFolder(string folder, CancellationToken token)
         {
+            if (token.IsCancellationRequested)
+            {
+                throw new TaskCanceledException("remove folder request on already canceled task");
+            }
+
             try
             {
                 Directory.Delete(folder, true);
@@ -87,8 +118,13 @@ namespace SlideCrafting.Utils
         }
 
         // https://docs.microsoft.com/en-us/dotnet/standard/io/how-to-copy-directories
-        public static void CopyFolder(string sourceDirName, string destDirName, bool copySubDirs = true)
+        public static void CopyFolder(string sourceDirName, string destDirName, CancellationToken token, bool copySubDirs = true)
         {
+            if (token.IsCancellationRequested)
+            {
+                throw new TaskCanceledException("copy files on already canceled task");
+            }
+
             // Get the subdirectories for the specified directory.
             DirectoryInfo dir = new DirectoryInfo(sourceDirName);
 
@@ -109,7 +145,7 @@ namespace SlideCrafting.Utils
             foreach (FileInfo file in files)
             {
                 string tempPath = Path.Combine(destDirName, file.Name);
-                file.CopyTo(tempPath, false);
+                file.CopyTo(tempPath, true);
             }
 
             // If copying subdirectories, copy them and their contents to new location.
@@ -118,36 +154,54 @@ namespace SlideCrafting.Utils
                 foreach (DirectoryInfo subdir in dirs)
                 {
                     string tempPath = Path.Combine(destDirName, subdir.Name);
-                    CopyFolder(subdir.FullName, tempPath, copySubDirs);
+                    CopyFolder(subdir.FullName, tempPath, token, copySubDirs);
                 }
             }
         }
 
         public static List<string> GetFilesOfFolder(string workFolder, string extensionFilter = null)
         {
+            if (!Directory.Exists(workFolder))
+            {
+                return new List<string>();
+            }
+
             var list = Directory.EnumerateFiles(workFolder)
-                .Where(x => extensionFilter == null || Path.GetExtension(x) == extensionFilter)
+                .Where(x => extensionFilter == null || x.EndsWith(extensionFilter))
                 .ToList();
             list.Sort();
             return list;
         }
 
-        public static List<string> GetInputFilesRecursive(string file, string extension, string settingKey)
+        public static List<string> GetInputFilesRecursive(string file, string baseFolder, string extension, string settingKey)
         {
+            var returnFiles = new List<string>();
+            
             // https://github.com/aaubry/YamlDotNet
             var yaml = new YamlStream();
-            yaml.Load(new StreamReader(file));
+            if (Path.IsPathFullyQualified(file))
+            {
+                yaml.Load(new StreamReader(file));
+            }
+            else
+            {
+                yaml.Load(new StreamReader(Path.Combine(baseFolder, file)));
+            }
 
             var root = (YamlMappingNode)yaml.Documents[0].RootNode;
+            if (!root.Children.ContainsKey(settingKey))
+            {
+                return returnFiles;
+            }
+
             YamlSequenceNode node = (YamlSequenceNode)root.Children[settingKey];
             var inputFiles = node.Children.Select(x => x.ToString());
 
-            var returnFiles = new List<string>();
             foreach (var inputFile in inputFiles)
             {
-                if (Path.GetExtension(inputFile) == extension)
+                if (inputFile.EndsWith(extension))
                 {
-                    returnFiles.AddRange(GetInputFilesRecursive(inputFile, extension, settingKey));
+                    returnFiles.AddRange(GetInputFilesRecursive(inputFile, baseFolder, extension, settingKey));
                 }
                 else
                 {
